Allow code to build for Swift compilers <5.5
Interchangable use of CGFloat and Double isn't supported until Swift 5.5:

https://github.com/apple/swift-evolution/blob/main/proposals/0307-allow-interchangeable-use-of-double-cgfloat-types.md

Addes #if swift(<5.5) conditional to allow compiling in older Swift versions.

Note: Only ran the tests on macOS 11.5.1.
All Test+Float tests pass in Swift 5.4.2 with this change, and it compiles in both 5.4.2 and 5.5, but there's a test assertion failure in Swift 5.5 (release version with Xcode 13.0).  That's a separate issue.